# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.11.0",
    "title": "SiSMoChat API",
    "description": "API for SiSMoChat - A secure chat application focused on child safety with parental control."
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT token. Format: \"Bearer {token}\""
    }
  },
  "paths": {
...
    "/parents/": {
      "post": {
        "tags": [
          "Messages"
        ],
        "summary": "Send a message",
        "description": "Send an encrypted message to another user.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Message sent, returns messageID"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/users/": {
      "post": {
        "tags": [
          "Messages"
        ],
        "summary": "Send a message",
        "description": "Send an encrypted message to another user.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Message sent, returns messageID"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/messages/": {
      "post": {
        "tags": [
          "Messages"
        ],
        "summary": "Send a message",
        "description": "Send an encrypted message to another user.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Message sent, returns messageID"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/connections/": {
      "get": {
        "tags": [
          "Status"
        ],
        "summary": "Get server status message",
        "description": "Returns an optional message for client display at startup (e.g. maintenance notices).",
        "responses": {
          "200": {
            "description": "Status message (null if none)"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/status/": {
      "get": {
        "tags": [
          "Status"
        ],
        "summary": "Get server status message",
        "description": "Returns an optional message for client display at startup (e.g. maintenance notices).",
        "responses": {
          "200": {
            "description": "Status message (null if none)"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Message": {
      "type": "object",
      "properties": {
        "to": {
          "type": "string",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        },
        "message": {
          "type": "string",
          "example": "encrypted message content"
        }
      }
    },
...
  }
}